# ism_swagger_RecyclEasy_218598.yaml
swagger: "2.0"
info:
  description: "The 'RecyclEasy' is a web application, that helps people find places near them, where they can dispose recyclabe garbage. As well as, publish advertisements related to recyclable garbage managment and disposal."
  version: "1.0.0"
  title: "RecycleEasy"
schemes:
  - "https"
  - "http"
paths:
...
  /place/create:
    post:
      tags:
        - "place"
      summary: "Create place"
      description: ""
      operationId: "createPlace"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - in: "body"
          name: "body"
          description: "Place creation payload"
          required: true
          schema:
            $ref: "#/definitions/PlaceCreationData"
      responses:
        200:
          description: "Place has been created successfully."
        400:
          description: "Some of the required fields are missing."
        403:
          description: "Only admin has access to this operation."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
  /place/update/{placeId}:
    put:
      tags:
        - "place"
      summary: "Update place"
      description: ""
      operationId: "updatePlace"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - in: "body"
          name: "body"
          description: "Place update payload"
          required: true
          schema:
            $ref: "#/definitions/PlaceUpdateData"
        - name: "placeId"
          in: "path"
          description: "ID of place to update"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Place has been updated successfully."
        400:
          description: "User not sent name or place"
        403:
          description: "Only admin has access to this operation."
        404:
          description: "Place doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
  /place/delete/{placeId}:
    delete:
      tags:
        - "place"
      summary: "Delete place"
      description: ""
      operationId: "deletePlace"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "placeId"
          in: "path"
          description: "ID of place to reove"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Place has been deleted."
        403:
          description: "Only admin has access to this operation.."
        404:
          description: "Place doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
...
  /place/getDetails/{placeId}:
    get:
      tags:
        - "place"
      summary: "Get place details"
      description: ""
      operationId: "placeDetails"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "placeId"
          in: "path"
          description: "ID of place"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Details of place with given Id"
          schema:
            $ref: "#/definitions/PlaceDetails"
        404:
          description: "Place doesnt exist."
  /place/filter:
    post:
      tags:
        - "place"
      summary: "Get places by specified filters"
      description: ""
      operationId: "getFilteredPlaces"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Place update payload"
          required: true
          schema:
            $ref: "#/definitions/PlaceFilterData"
      responses:
        200:
          description: "Returns objects meeting specified filters."
          schema:
            $ref: "#/definitions/Place"
  /place/suggest:
    post:
      tags:
        - "place"
      summary: "Suggest new place"
      description: ""
      operationId: "sendSuggestion"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Place creation payload"
          required: true
          schema:
            $ref: "#/definitions/PlaceCreationData"
      responses:
        200:
          description: "Suggestion sent successfully."
...
  /place/approve/{suggestionId}:
    post:
      tags:
        - "admin"
      summary: "Approve suggestion with given ID."
      description: ""
      operationId: "approveSuggestion"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "suggestionId"
          in: "path"
          description: "ID of place suggestion"
          required: true
          type: number
      responses:
        200:
          description: "Place approved successfully."
        403:
          description: "Only admin has access to this operation."
        404:
          description: "Suggestion doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
  /place/reject/{suggestionId}:
    post:
      tags:
        - "admin"
      summary: "Reject place suggestion with given ID"
      description: ""
      operationId: "rejectSuggestion"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "suggestionId"
          in: "path"
          description: "ID of place suggestion"
          required: true
          type: number
      responses:
        200:
          description: "Place rejected successfully."
        403:
          description: "Only admin has access to this operation."
        404:
          description: "Suggestion doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
...
securityDefinitions:
  recycleasy_auth:
    type: "oauth2"
    authorizationUrl: "http://find.your.team.com/login"
    flow: "implicit"
    scopes:
      write: "modify data"
      read: "read data"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Place:
    type: object
    properties:
      id:
        type: number
      name:
        type: string
      address:
        type: string
      locationX:
        type: number
      locationY:
        type: number
  PlaceDetails:
    type: object
    properties:
      id:
        type: number
      name:
        type: string
      address:
        type: string
      locationX:
        type: number
      locationY:
        type: number
      opensAt:
        type: string
      closesAt:
        type: string
      phone:
        type: string
      email:
        type: string
  PlaceCreationData:
    type: object
    required:
      - name
      - address
      - locationX
      - locationY
      - phone
    properties:
      name:
        type: string
      address:
        type: string
      locationX:
        type: string
      locationY:
        type: string
      phone:
        type: string
      opensAt:
        type: string
      closesAt:
        type: string
      garbageTypes:
        type: array
        items:
          type: string
  PlaceUpdateData:
    type: object
    required:
      - name
      - address
      - locationX
      - locationY
      - phone
    properties:
      name:
        type: string
      address:
        type: string
      locationX:
        type: string
      locationY:
        type: string
      phone:
        type: string
      opensAt:
        type: string
      closesAt:
        type: string
      garbageTypes:
        type: array
        items:
          type: string
  PlaceFilterData:
    type: object
    properties:
      city:
        type: string
      garbagetypes:
        type: array
        items:
          type: string
...